+++ gcp_integration_test.go
@@ -20,6 +20,9 @@ func TestCreateGCPIntegration(t *testing.T) {
 	})
 	assert.NoError(t, err, "Unexpected error creating integration")
 	assert.Equal(t, "string", result.Name, "Name does not match")
+	assert.Equal(t, integration.FiveMinutely, *result.PollRate, "PollRate does not match")
+	assert.Equal(t, true, result.UseMetricSourceProjectForQuota, "UseMetricSourceProjectForQuota does not match")
+	assert.Equal(t, int64(300000), result.PollRateMs, "PollRateMs does not match")
 }
 
 func TestGetGCPIntegration(t *testing.T) {