+++ integration/model_aws_cloud_watch_integration.go
@@ -66,4 +66,6 @@ type AwsCloudWatchIntegration struct {
 	LogsSyncState string `json:"logsSyncState,omitempty"`
 	// Name of an existing access token (org token) in your Observability Cloud organization. Used to track usage metrics.
 	NamedToken string `json:"namedToken,omitempty"`
+	// If `true`, this property indicates that SignalFx should sync metrics and metadata from custom AWS namespaces only.
+	SyncCustomNamespacesOnly bool `json:"syncCustomNamespacesOnly,omitempty"`
 }