+++ integration/model_aws_service.go
@@ -9,188 +9,188 @@
 
 package integration
 
-// AwsService : An AWS service that you want SignalFx to collect data from. SignalFx supports the following AWS services:<br>   * AWS/ApiGateway   * AWS/AppStream   * AWS/AutoScaling   * AWS/Backup   * AWS/Billing   * AWS/CloudFront   * AWS/CloudSearch   * AWS/Events   * AWS/Logs   * AWS/Connect   * AWS/DMS   * AWS/DX   * AWS/DynamoDB   * AWS/EC2   * AWS/EC2Spot   * AWS/ECS   * AWS/ElasticBeanstalk   * AWS/EBS   * AWS/EFS   * AWS/ELB   * AWS/ApplicationELB   * AWS/NetworkELB   * AWS/ElasticTranscoder   * AWS/ElastiCache   * AWS/ES   * AWS/ElasticMapReduce   * AWS/GameLift   * AWS/Inspector   * AWS/IoT   * AWS/KMS   * AWS/KinesisAnalytics   * AWS/Firehose   * AWS/Kinesis   * AWS/KinesisVideo   * AWS/Lambda   * AWS/Lex   * AWS/ML   * AWS/OpsWorks   * AWS/Polly   * AWS/Redshift   * AWS/RDS   * AWS/Route53   * AWS/SageMaker   * AWS/DDoSProtection   * AWS/SES   * AWS/SNS   * AWS/SQS   * AWS/S3   * AWS/SWF   * AWS/States   * AWS/StorageGateway   * AWS/Translate   * AWS/NATGateway   * AWS/VPN   * AWS/WAFV2   * WAF   * AWS/WorkSpaces   * CWAgent
 type AwsService string
 
-// List of AWSService
 const (
+	AWSACP_PRIVATE_CA            AwsService = "AWS/ACMPrivateCA"
+	AWSAMAZON_MQ                 AwsService = "AWS/AmazonMQ"
 	AWSAPI_GATEWAY               AwsService = "AWS/ApiGateway"
+	AWSAPPLICATION_ELB           AwsService = "AWS/ApplicationELB"
 	AWSAPP_STREAM                AwsService = "AWS/AppStream"
 	AWSATHENA                    AwsService = "AWS/Athena"
 	AWSAUTO_SCALING              AwsService = "AWS/AutoScaling"
 	AWSBACKUP                    AwsService = "AWS/Backup"
 	AWSBILLING                   AwsService = "AWS/Billing"
-	AWSACP_PRIVATE_CA            AwsService = "AWS/ACMPrivateCA"
 	AWSCERT_MANAGER              AwsService = "AWS/CertificateManager"
 	AWSCLOUD_FRONT               AwsService = "AWS/CloudFront"
 	AWSCLOUD_HSM                 AwsService = "AWS/CloudHSM"
 	AWSCLOUD_SEARCH              AwsService = "AWS/CloudSearch"
-	AWSEVENTS                    AwsService = "AWS/Events"
-	AWSLOGS                      AwsService = "AWS/Logs"
 	AWSCODEBUILD                 AwsService = "AWS/CodeBuild"
 	AWSCOGNITO                   AwsService = "AWS/Cognito"
 	AWSCONNECT                   AwsService = "AWS/Connect"
+	AWSD_DO_S_PROTECTION         AwsService = "AWS/DDoSProtection"
 	AWSDMS                       AwsService = "AWS/DMS"
-	AWSDX                        AwsService = "AWS/DX"
 	AWSDOCDB                     AwsService = "AWS/DocDB"
+	AWSDX                        AwsService = "AWS/DX"
 	AWSDYNAMO_DB                 AwsService = "AWS/DynamoDB"
+	AWSEBS                       AwsService = "AWS/EBS"
 	AWSEC2                       AwsService = "AWS/EC2"
 	AWSEC2_SPOT                  AwsService = "AWS/EC2Spot"
 	AWSECS                       AwsService = "AWS/ECS"
-	AWSELASTIC_BEANSTALK         AwsService = "AWS/ElasticBeanstalk"
-	AWSELASTIC_INTERFACE         AwsService = "AWS/ElasticInterface"
-	AWSEBS                       AwsService = "AWS/EBS"
 	AWSEFS                       AwsService = "AWS/EFS"
 	AWSEKS                       AwsService = "AWS/EKS"
-	AWSELB                       AwsService = "AWS/ELB"
-	AWSAPPLICATION_ELB           AwsService = "AWS/ApplicationELB"
-	AWSNETWORK_ELB               AwsService = "AWS/NetworkELB"
-	AWSELASTIC_TRANSCODER        AwsService = "AWS/ElasticTranscoder"
 	AWSELASTI_CACHE              AwsService = "AWS/ElastiCache"
-	AWSES                        AwsService = "AWS/ES"
+	AWSELASTIC_BEANSTALK         AwsService = "AWS/ElasticBeanstalk"
+	AWSELASTIC_INTERFACE         AwsService = "AWS/ElasticInterface"
 	AWSELASTIC_MAP_REDUCE        AwsService = "AWS/ElasticMapReduce"
+	AWSELASTIC_TRANSCODER        AwsService = "AWS/ElasticTranscoder"
+	AWSELB                       AwsService = "AWS/ELB"
+	AWSES                        AwsService = "AWS/ES"
+	AWSEVENTS                    AwsService = "AWS/Events"
+	AWSFIREHOSE                  AwsService = "AWS/Firehose"
 	AWSFSX                       AwsService = "AWS/FSx"
 	AWSGAME_LIFT                 AwsService = "AWS/GameLift"
-	GLUE                         AwsService = "Glue"
 	AWSINSPECTOR                 AwsService = "AWS/Inspector"
 	AWSIO_T                      AwsService = "AWS/IoT"
 	AWSIO_T_ANALYTICS            AwsService = "AWS/IoTAnalytics"
 	AWSKAFKA                     AwsService = "AWS/Kafka"
-	AWSKMS                       AwsService = "AWS/KMS"
-	AWSKINESIS_ANALYTICS         AwsService = "AWS/KinesisAnalytics"
-	AWSFIREHOSE                  AwsService = "AWS/Firehose"
 	AWSKINESIS                   AwsService = "AWS/Kinesis"
+	AWSKINESIS_ANALYTICS         AwsService = "AWS/KinesisAnalytics"
 	AWSKINESIS_VIDEO             AwsService = "AWS/KinesisVideo"
+	AWSKMS                       AwsService = "AWS/KMS"
 	AWSLAMBDA                    AwsService = "AWS/Lambda"
 	AWSLEX                       AwsService = "AWS/Lex"
+	AWSLOGS                      AwsService = "AWS/Logs"
 	AWSMEDIA_CONNECT             AwsService = "AWS/MediaConnect"
 	AWSMEDIA_CONVERT             AwsService = "AWS/MediaConvert"
 	AWSMEDIA_PACKAGE             AwsService = "AWS/MediaPackage"
 	AWSMEDIA_TAILOR              AwsService = "AWS/MediaTailor"
 	AWSML                        AwsService = "AWS/ML"
-	AWSAMAZON_MQ                 AwsService = "AWS/AmazonMQ"
+	AWSNAT_GATEWAY               AwsService = "AWS/NATGateway"
+	AWSNEPTUNE                   AwsService = "AWS/Neptune"
+	AWSNETWORK_ELB               AwsService = "AWS/NetworkELB"
 	AWSOPS_WORKS                 AwsService = "AWS/OpsWorks"
 	AWSPOLLY                     AwsService = "AWS/Polly"
-	AWSREDSHIFT                  AwsService = "AWS/Redshift"
 	AWSRDS                       AwsService = "AWS/RDS"
+	AWSREDSHIFT                  AwsService = "AWS/Redshift"
 	AWSROBOMAKER                 AwsService = "AWS/Robomaker"
 	AWSROUTE53                   AwsService = "AWS/Route53"
+	AWSS3                        AwsService = "AWS/S3"
+	AWSS3_STORAGE_LENS           AwsService = "AWS/S3/Storage-Lens"
 	AWSSAGE_MAKER                AwsService = "AWS/SageMaker"
+	AWSSAGE_MAKER_ENDPOINTS      AwsService = "aws/sagemaker/Endpoints"
 	AWSSAGE_MAKER_TRAINING_JOBS  AwsService = "aws/sagemaker/TrainingJobs"
 	AWSSAGE_MAKER_TRANSFORM_JOBS AwsService = "aws/sagemaker/TransformJobs"
-	AWSSAGE_MAKER_ENDPOINTS      AwsService = "aws/sagemaker/Endpoints"
 	AWSSDK_METRICS               AwsService = "AWS/SDKMetrics"
-	AWSD_DO_S_PROTECTION         AwsService = "AWS/DDoSProtection"
 	AWSSES                       AwsService = "AWS/SES"
 	AWSSNS                       AwsService = "AWS/SNS"
 	AWSSQS                       AwsService = "AWS/SQS"
-	AWSS3                        AwsService = "AWS/S3"
-	AWSSWF                       AwsService = "AWS/SWF"
 	AWSSTATES                    AwsService = "AWS/States"
 	AWSSTORAGE_GATEWAY           AwsService = "AWS/StorageGateway"
+	AWSSWF                       AwsService = "AWS/SWF"
 	AWSTEXTRACT                  AwsService = "AWS/Textract"
 	AWSTHINGS_GRAPH              AwsService = "AWS/ThingsGraph"
 	AWSTRANSLATE                 AwsService = "AWS/Translate"
 	AWSTRUSTED_ADVISOR           AwsService = "AWS/TrustedAdvisor"
-	AWSNAT_GATEWAY               AwsService = "AWS/NATGateway"
 	AWSVPN                       AwsService = "AWS/VPN"
 	AWSWAFV2                     AwsService = "AWS/WAFV2"
-	WAF                          AwsService = "WAF"
 	AWSWORK_MAIL                 AwsService = "AWS/WorkMail"
 	AWSWORK_SPACES               AwsService = "AWS/WorkSpaces"
-	AWSNEPTUNE                   AwsService = "AWS/Neptune"
+	CWAGENT                      AwsService = "CWAgent"
+	GLUE                         AwsService = "Glue"
 	AWSMEDIA_LIVE                AwsService = "MediaLive"
 	AWS_SYSTEM_LINUX             AwsService = "System/Linux"
-	CWAGENT                      AwsService = "CWAgent"
+	WAF                          AwsService = "WAF"
 )
 
 var AWSServiceNames = map[string]AwsService{
+	"AWS/ACMPrivateCA":            AWSACP_PRIVATE_CA,
+	"AWS/AmazonMQ":                AWSAMAZON_MQ,
 	"AWS/ApiGateway":              AWSAPI_GATEWAY,
+	"AWS/ApplicationELB":          AWSAPPLICATION_ELB,
 	"AWS/AppStream":               AWSAPP_STREAM,
 	"AWS/Athena":                  AWSATHENA,
 	"AWS/AutoScaling":             AWSAUTO_SCALING,
 	"AWS/Backup":                  AWSBACKUP,
 	"AWS/Billing":                 AWSBILLING,
-	"AWS/ACMPrivateCA":            AWSACP_PRIVATE_CA,
 	"AWS/CertificateManager":      AWSCERT_MANAGER,
 	"AWS/CloudFront":              AWSCLOUD_FRONT,
 	"AWS/CloudHSM":                AWSCLOUD_HSM,
 	"AWS/CloudSearch":             AWSCLOUD_SEARCH,
-	"AWS/Events":                  AWSEVENTS,
-	"AWS/Logs":                    AWSLOGS,
 	"AWS/CodeBuild":               AWSCODEBUILD,
 	"AWS/Cognito":                 AWSCOGNITO,
 	"AWS/Connect":                 AWSCONNECT,
+	"AWS/DDoSProtection":          AWSD_DO_S_PROTECTION,
 	"AWS/DMS":                     AWSDMS,
-	"AWS/DX":                      AWSDX,
 	"AWS/DocDB":                   AWSDOCDB,
+	"AWS/DX":                      AWSDX,
 	"AWS/DynamoDB":                AWSDYNAMO_DB,
+	"AWS/EBS":                     AWSEBS,
 	"AWS/EC2":                     AWSEC2,
 	"AWS/EC2Spot":                 AWSEC2_SPOT,
 	"AWS/ECS":                     AWSECS,
-	"AWS/ElasticBeanstalk":        AWSELASTIC_BEANSTALK,
-	"AWS/ElasticInterface":        AWSELASTIC_INTERFACE,
-	"AWS/EBS":                     AWSEBS,
 	"AWS/EFS":                     AWSEFS,
 	"AWS/EKS":                     AWSEKS,
-	"AWS/ELB":                     AWSELB,
-	"AWS/ApplicationELB":          AWSAPPLICATION_ELB,
-	"AWS/NetworkELB":              AWSNETWORK_ELB,
-	"AWS/ElasticTranscoder":       AWSELASTIC_TRANSCODER,
 	"AWS/ElastiCache":             AWSELASTI_CACHE,
-	"AWS/ES":                      AWSES,
+	"AWS/ElasticBeanstalk":        AWSELASTIC_BEANSTALK,
+	"AWS/ElasticInterface":        AWSELASTIC_INTERFACE,
 	"AWS/ElasticMapReduce":        AWSELASTIC_MAP_REDUCE,
+	"AWS/ElasticTranscoder":       AWSELASTIC_TRANSCODER,
+	"AWS/ELB":                     AWSELB,
+	"AWS/ES":                      AWSES,
+	"AWS/Events":                  AWSEVENTS,
+	"AWS/Firehose":                AWSFIREHOSE,
 	"AWS/FSx":                     AWSFSX,
 	"AWS/GameLift":                AWSGAME_LIFT,
-	"Glue":                        GLUE,
 	"AWS/Inspector":               AWSINSPECTOR,
 	"AWS/IoT":                     AWSIO_T,
 	"AWS/IoTAnalytics":            AWSIO_T_ANALYTICS,
 	"AWS/Kafka":                   AWSKAFKA,
-	"AWS/KMS":                     AWSKMS,
-	"AWS/KinesisAnalytics":        AWSKINESIS_ANALYTICS,
-	"AWS/Firehose":                AWSFIREHOSE,
 	"AWS/Kinesis":                 AWSKINESIS,
+	"AWS/KinesisAnalytics":        AWSKINESIS_ANALYTICS,
 	"AWS/KinesisVideo":            AWSKINESIS_VIDEO,
+	"AWS/KMS":                     AWSKMS,
 	"AWS/Lambda":                  AWSLAMBDA,
 	"AWS/Lex":                     AWSLEX,
+	"AWS/Logs":                    AWSLOGS,
 	"AWS/MediaConnect":            AWSMEDIA_CONNECT,
 	"AWS/MediaConvert":            AWSMEDIA_CONVERT,
 	"AWS/MediaPackage":            AWSMEDIA_PACKAGE,
 	"AWS/MediaTailor":             AWSMEDIA_TAILOR,
 	"AWS/ML":                      AWSML,
-	"AWS/AmazonMQ":                AWSAMAZON_MQ,
+	"AWS/NATGateway":              AWSNAT_GATEWAY,
+	"AWS/Neptune":                 AWSNEPTUNE,
+	"AWS/NetworkELB":              AWSNETWORK_ELB,
 	"AWS/OpsWorks":                AWSOPS_WORKS,
 	"AWS/Polly":                   AWSPOLLY,
-	"AWS/Redshift":                AWSREDSHIFT,
 	"AWS/RDS":                     AWSRDS,
+	"AWS/Redshift":                AWSREDSHIFT,
 	"AWS/Robomaker":               AWSROBOMAKER,
 	"AWS/Route53":                 AWSROUTE53,
+	"AWS/S3":                      AWSS3,
+	"AWS/S3/Storage-Lens":         AWSS3_STORAGE_LENS,
 	"AWS/SageMaker":               AWSSAGE_MAKER,
+	"aws/sagemaker/Endpoints":     AWSSAGE_MAKER_ENDPOINTS,
 	"aws/sagemaker/TrainingJobs":  AWSSAGE_MAKER_TRAINING_JOBS,
 	"aws/sagemaker/TransformJobs": AWSSAGE_MAKER_TRANSFORM_JOBS,
-	"aws/sagemaker/Endpoints":     AWSSAGE_MAKER_ENDPOINTS,
 	"AWS/SDKMetrics":              AWSSDK_METRICS,
-	"AWS/DDoSProtection":          AWSD_DO_S_PROTECTION,
 	"AWS/SES":                     AWSSES,
 	"AWS/SNS":                     AWSSNS,
 	"AWS/SQS":                     AWSSQS,
-	"AWS/S3":                      AWSS3,
-	"AWS/SWF":                     AWSSWF,
 	"AWS/States":                  AWSSTATES,
 	"AWS/StorageGateway":          AWSSTORAGE_GATEWAY,
+	"AWS/SWF":                     AWSSWF,
 	"AWS/Textract":                AWSTEXTRACT,
 	"AWS/ThingsGraph":             AWSTHINGS_GRAPH,
 	"AWS/Translate":               AWSTRANSLATE,
 	"AWS/TrustedAdvisor":          AWSTRUSTED_ADVISOR,
-	"AWS/NATGateway":              AWSNAT_GATEWAY,
 	"AWS/VPN":                     AWSVPN,
 	"AWS/WAFV2":                   AWSWAFV2,
-	"WAF":                         WAF,
 	"AWS/WorkMail":                AWSWORK_MAIL,
 	"AWS/WorkSpaces":              AWSWORK_SPACES,
-	"AWS/Neptune":                 AWSNEPTUNE,
+	"CWAgent":                     CWAGENT,
+	"Glue":                        GLUE,
 	"MediaLive":                   AWSMEDIA_LIVE,
 	"System/Linux":                AWS_SYSTEM_LINUX,
-	"CWAgent":                     CWAGENT,
+	"WAF":                         WAF,
 }